+++ callng/utilities.lua
@@ -13,14 +13,13 @@ syslog = require("posix.syslog")
 json = require("json")
 redis = require("redis")
 require("callng.configuration")
-log = require("callng.nglog")
-
-log.stacks = {
+nglog = require("callng.nglog")
+nglog.stacks = {
     console =   LOGSTACK_CONSOLE,
     file    =   LOGSTACK_FILE,
     syslog  =   LOGSTACK_SYSLOG
 }
-log.level, log.host, log.name = LOGLEVEL, NODEID, 'libresbc'
+nglog.level, nglog.host, nglog.name = LOGLEVEL, NODEID, 'libresbc'
 
 random = math.random
 unpack = _G.unpack or table.unpack
@@ -152,6 +151,23 @@ function jsonhash(data)
         end
     end
     return data
+end
+
+----------------------------------------------------------------------------
+--- switch log
+log = {
+    debug    = function(text, ...) freeswitch.consoleLog("debug",   string.format(text.."\n", ...)) end,
+    info     = function(text, ...) freeswitch.consoleLog("info",    string.format(text.."\n", ...)) end,
+    notice   = function(text, ...) freeswitch.consoleLog("notice",  string.format(text.."\n", ...)) end,
+    warning  = function(text, ...) freeswitch.consoleLog("warning", string.format(text.."\n", ...)) end,
+    error    = function(text, ...) freeswitch.consoleLog("err",     string.format(text.."\n", ...)) end,
+    critical = function(text, ...) freeswitch.consoleLog("crit",    string.format(text.."\n", ...)) end,
+    alert    = function(text, ...) freeswitch.consoleLog("alert",   string.format(text.."\n", ...)) end
+}
+
+-- back to nglog
+if LOGSTACK_CONSOLE or LOGSTACK_FILE or LOGSTACK_SYSLOG then
+    log = nglog
 end
 
 ----------------------------------------------------------------------------